% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova.R
\name{ano2}
\alias{ano2}
\title{A method function that calculates the ANOVA table given a vector of responses
an a vector of categorical covariate observation}
\usage{
ano2(response, covariates)
}
\arguments{
\item{response}{vector of responses}

\item{covariates}{vector of categorical covariate observation}
}
\value{
factor and residual degrees of freedom, between group sum of squares,
total sum of squares, mean squared errors, F-statistic, p-value
}
\description{
A method function that calculates the ANOVA table given a vector of responses
an a vector of categorical covariate observation
}
