% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConfTtest.R
\name{ConfTtest}
\alias{ConfTtest}
\title{A function that calculates t-test for each coefficient}
\usage{
ConfTtest(object)
}
\arguments{
\item{object}{an object of LReg}
}
\value{
estimates, standard error, and confidence intervals of coefficients
}
\description{
A function that calculates t-test for each coefficient
}
\examples{
ConfTtest(LReg(Species ~ Area +Elevation+ Nearest+ Scruz+ Adjacent, mydata1))
}
