% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionlist.R
\name{ContrastTest}
\alias{ContrastTest}
\title{Calculate F statistic for contrast comparisons}
\usage{
ContrastTest(object, R, alpha = 0.05)
}
\arguments{
\item{object}{an object of LReg}

\item{R}{a matrix of contrast}

\item{alpha}{level of significance (default is 0.05)}
}
\value{
F statistic, quantile of F distribution at the level of 1-alpha, and significance status
}
\description{
Calculate F statistic for contrast comparisons
}
