% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{Diagnostic_Plots}
\alias{Diagnostic_Plots}
\title{Return a list of diagnostic plots}
\usage{
Diagnostic_Plots(object)
}
\arguments{
\item{object}{an object of LReg}
}
\value{
list of diagnostic plots
}
\description{
Return a list of diagnostic plots
}
