% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelcomp.R
\name{ModelComparison}
\alias{ModelComparison}
\title{Return the f statistic for the comparison of a larger vs a smaller model, null hypothesis is small model is correct}
\usage{
ModelComparison(large_model, small_model)
}
\arguments{
\item{large_model}{a formula}

\item{small_model}{a formula, y must be same as large_model and all variables must also be in large_model}
}
\value{
f statistic
}
\description{
Return the f statistic for the comparison of a larger vs a smaller model, null hypothesis is small model is correct
}
