% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mydata1.R
\docType{data}
\name{mydata1}
\alias{mydata1}
\title{Species diversity on the Galapagos Islands}
\format{
There are 30 Galapagos islands and 6 variables in the dataset.
The relationship between the number of plant species and several
geographic variables is of interest. The original dataset contained
several missing values which have been filled for convenience.
\describe{
\item{Species}{The number of species found on the island}
\item{Area}{The area of the island (km^2)}
\item{Elevation}{The highest elevation of the island (m)}
\item{Nearest}{The distance from the nearest island (km)}
\item{Scruz}{The distance from Santa Cruz Island (km)}
\item{Adjacent}{The area of the adjacent island (km2)}
}
}
\source{
{M. P. Johnson and P. H. Raven (1973) "Species number and endemism: The Galapagos Archipelago revisited" Science, 179, 893-895}
}
\usage{
mydata1
}
\description{
Report the number of species found on the various Galápagos Islands
}
\keyword{datasets}
