% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mydatalist.R
\docType{data}
\name{mydata2}
\alias{mydata2}
\title{The odor of household products}
\format{
There are 15 observations and 3 factors, temperature, gas/liquid ratio and packing height,
in the dataset. Each factor has 3 levels (-1, 0, 1).

\describe{
\item{odor}{The level of unpleasant odor}
\item{temp}{the temperature factor with 3 levels}
\item{gas}{The gas/liquid ratio categorized into 3 levels}
\item{pack}{The packing height categorized into 3 levels}
}
}
\source{
{Baker FD, Bargmann RE. Orthogonal central composite designs of the third order in the evaluation of sensitivity and plant growth simulation models. Journal of the American Statistical Association. 1985 Sep 1;80(391):574-9.}
}
\usage{
mydata2
}
\description{
Report the level of unpleasant odor explained by temperature, gas/liquid ratio and packing height.
}
\keyword{datasets}
