% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LReg.R
\name{print.tconf}
\alias{print.tconf}
\title{A method function that calculates t-test for each coefficient}
\usage{
\method{print}{tconf}(x, ...)
}
\arguments{
\item{x}{an object of LReg}

\item{...}{redundant argument}
}
\value{
estimates, standard error, and confidence intervals of coefficients
}
\description{
A method function that calculates t-test for each coefficient
}
