% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reaov-package.R
\docType{package}
\name{reaov-package}
\alias{reaov}
\alias{reaov-package}
\title{reaov: A project of MATH533: A R package for regression and ANOVA}
\description{
This package involves R functions for linear regression analysis and ANOVA. We try not to use the built-in functions but manual formulae in the lecture notes.
}
\author{
\strong{Maintainer}: Yanees Dobberstein \email{yanees.dobberstein@mail.mcgill.ca}

Authors:
\itemize{
  \item Nam-Anh Tran \email{nam-anh.tran@mail.mcgill.ca}
}

}
\keyword{internal}
