% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova.R
\name{Rsquared}
\alias{Rsquared}
\title{A function that calculated the R squared and adjusted R squared values of a
model}
\usage{
Rsquared(object)
}
\arguments{
\item{object}{an object of LReg}
}
\value{
R squared and adjusted R squared value of object
}
\description{
A function that calculated the R squared and adjusted R squared values of a
model
}
