% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionlist.R
\name{SubsetBetaTest}
\alias{SubsetBetaTest}
\title{Calculate F statistic to test sub-vector of coefficients}
\usage{
SubsetBetaTest(object, reduced = 1, alpha = 0.05)
}
\arguments{
\item{object}{an object of LReg}

\item{reduced}{a vector of numbers relating to the ordering of reduced predictors (The ordering depends on the formula in the object of LReg)}

\item{alpha}{level of significance (default is 0.05)}
}
\value{
F statistic,  quantile of F distribution at the level of 1-alpha, and significance status
}
\description{
Calculate F statistic to test sub-vector of coefficients
}
