% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionlist.R
\name{summary.CIresponse}
\alias{summary.CIresponse}
\title{Calculate the 95\% confidence interval for the mean response and future response}
\usage{
\method{summary}{CIresponse}(object, x_star = NULL, alpha = 0.05, ...)
}
\arguments{
\item{object}{an object of LReg}

\item{x_star}{a matrix of new observations}

\item{alpha}{level of significance (default is 0.05)}

\item{...}{redundant argument}
}
\value{
CI of mean response and future response
}
\description{
Calculate the 95\% confidence interval for the mean response and future response
}
