% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tconf.R
\name{summary.tconf}
\alias{summary.tconf}
\title{A method function that calculates t-test for each coefficient}
\usage{
\method{summary}{tconf}(object, ...)
}
\arguments{
\item{object}{an object of LReg}

\item{...}{redundant argument}
}
\value{
estimates, standard error, and confidence intervals of coefficients
}
\description{
A method function that calculates t-test for each coefficient
}
