% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LReg.R
\name{AddClass}
\alias{AddClass}
\title{A method function that modifies the class of LReg object}
\usage{
AddClass(object_of_LReg, myclass = NULL)
}
\arguments{
\item{object_of_LReg}{an object of LReg}

\item{myclass}{a string variable. This is the new class you want to add to the object}
}
\value{
the LReg object with the new class
}
\description{
A method function that modifies the class of LReg object
}
