% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionlist.R
\name{summary.JCRpars}
\alias{summary.JCRpars}
\title{Calculate joint hypothesis testing for the model parameters}
\usage{
\method{summary}{JCRpars}(object, beta = rep(0, object$p + 1), alpha = 0.05, ...)
}
\arguments{
\item{object}{an object of LReg}

\item{beta}{the values of testing}

\item{alpha}{level of significance (default is 0.05)}

\item{...}{redundant argument}
}
\value{
F statistic, lower and upper bound of F distribution at the level of alpha, and significance status
}
\description{
Calculate joint hypothesis testing for the model parameters
}
